Start on Proper Card Placement + Refactoring
You can now move individual cards around and they will snap to their given pile if they can be placed, or go back to where they came from if they can't be placed in the new pile. It's still buggy because it does not work with multiple cards, but that is something I shall work on next time.

Did some refactoring, namely changing all the board_pile name to depot, as that is their proper name, as well as adding some type aliases. Still likely more to go but this is good enough for now and I've done a lot so far.